requirements.txt
Django==2.2.2
django-embed-video==1.3.3
django-heroku==0.3.1
django-import-export==2.3.0
gunicorn==20.0.4
Jinja2==2.11.2
virtualenv==20.0.21
